Refactor road pathfinding into MapPathing.kt
